Further work on extracting Extracting files is threaded with goroutines. Extracting also sets the proper UID and GUID Made recursive children getting threaded with goroutines Decided I will allow wildcards in paths. Hasn't been implemented though. Fixed issues with ExtendedDirectory reading (I was using a uint16 instead of a uint32) I couldn't test basically any of this. somehow a for loop isn't incrementing. For seemingly no reason.

+++ internal/directory/directory.go
@@ -52,7 +52,7 @@ type Directory struct {
 }
 
 //NewDirectory reads the directory from rdr
-func NewDirectory(base io.Reader, size uint16) (*Directory, error) {
+func NewDirectory(base io.Reader, size uint32) (*Directory, error) {
 	var dir Directory
 	var err error
 	tmp := make([]byte, size)
